move pdp to cms-pages structure
