+++ projects/storefrontlib/src/cms-pages/index.ts
@@ -1,0 +1,2 @@
+export * from './product-details-page/product-details-page.component';
+export * from './product-details-page/product-details-page.module';
