+++ projects/storefrontlib/src/cms-pages/product-details-page/product-details-page.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'cx-product-details-page',
+  templateUrl: './product-details-page.component.html',
+})
+export class ProductDetailsPageComponent {}
