+++ projects/storefrontlib/src/lib/storefront.module.ts
@@ -1,11 +1,10 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
-
 import {
-  OccModule,
   AuthModule,
   ConfigModule,
   CxApiModule,
   I18nModule,
+  OccModule,
   PersonalizationModule,
   provideConfig,
   RoutingModule,
@@ -13,6 +12,7 @@ import {
   StateModule,
 } from '@spartacus/core';
 import { CmsLibModule } from '../cms-components/index';
+import { ProductDetailsPageModule } from '../cms-pages/product-details-page/product-details-page.module';
 import { CmsModule } from '../cms-structure/cms.module';
 import { CmsRouteModule } from '../cms-structure/routing/cms-route/cms-route.module';
 import { SuffixRoutesModule } from '../cms-structure/routing/suffix-routes/suffix-routes.module';
@@ -36,6 +36,8 @@ import { UiModule } from './ui/index';
     SmartEditModule.forRoot(),
     PersonalizationModule.forRoot(),
     I18nModule.forRoot(),
+    // pages
+    ProductDetailsPageModule,
   ],
   exports: [UiModule],
   providers: [...provideConfigFromMetaTags()],
