+++ projects/storefrontlib/src/lib/ui/pages/index.ts
@@ -1,6 +1,4 @@
-export * from './pages.module';
-export * from './product-page/product-page.component';
 export * from './cart-page/cart-page.component';
-export * from './order-confirmation-page/order-confirmation-page.component';
 export * from './cart-page/cart-page.module';
-export * from './product-page/product-page.module';
+export * from './order-confirmation-page/order-confirmation-page.component';
+export * from './pages.module';
