--- projects/storefrontlib/src/lib/ui/pages/product-page/product-page.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'cx-product-page',
-  templateUrl: './product-page.component.html',
-})
-export class ProductPageComponent {}
