--- projects/storefrontlib/src/lib/ui/pages/product-page/product-page.module.ts
@@ -1,30 +1,0 @@
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { RouterModule, Routes } from '@angular/router';
-import { ProductDetailsModule } from '../../../../cms-components/product/index';
-import { CmsPageGuard } from '../../../../cms-structure/guards/cms-page.guard';
-import { PageLayoutModule } from '../../../../cms-structure/page/index';
-import { ProductPageComponent } from './product-page.component';
-import { OutletRefModule } from '../../../../cms-structure/outlet/outlet-ref/outlet-ref.module';
-
-const routes: Routes = [
-  {
-    path: null,
-    canActivate: [CmsPageGuard],
-    component: ProductPageComponent,
-    data: { cxRoute: 'product' },
-  },
-];
-
-@NgModule({
-  imports: [
-    CommonModule,
-    RouterModule.forChild(routes),
-    ProductDetailsModule,
-    PageLayoutModule,
-    OutletRefModule,
-  ],
-  declarations: [ProductPageComponent],
-  exports: [ProductPageComponent],
-})
-export class ProductPageModule {}
